whisper-kokoro/requirements.txt
fastapi
uvicorn
openai-whisper
torch
python-multipart
kokoro
soundfile
numpy